+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [1.0.2] - 2019-08-16
 ### Added
 - Post reporting.
 - Project reporting.
@@ -12,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Fix google auth.
 - Improve UI.
+- Show terms of service on LoginScreen.
 
 ## [1.0.1] - 2019-08-14
 ### Changed
@@ -21,6 +24,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - CHANGELOG file.
 
-[Unreleased]: https://github.com/ccmikechen/Birdiy-app/compare/v1.0.1...HEAD
+[Unreleased]: https://github.com/ccmikechen/Birdiy-app/compare/v1.0.2...HEAD
+[1.0.2]: https://github.com/ccmikechen/Birdiy-app/releases/tag/v1.0.2
 [1.0.1]: https://github.com/ccmikechen/Birdiy-app/releases/tag/v1.0.1
 [1.0.0]: https://github.com/ccmikechen/Birdiy-app/releases/tag/v1.0.0